requirements/tests.txt
pytest~=3.0.0
pytest-selenium==1.7.0
pytest-variables==1.4
requests==2.12.4
